+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/LoRaDeviceAPIService.cs
@@ -27,7 +27,7 @@ namespace LoRaWan.NetworkServer
             this.serviceFacadeHttpClientProvider = serviceFacadeHttpClientProvider;
         }
 
-        public override async Task<ushort> NextFCntDownAsync(string devEUI, int fcntDown, int fcntUp, string gatewayId)
+        public override async Task<uint> NextFCntDownAsync(string devEUI, uint fcntDown, uint fcntUp, string gatewayId)
         {
             Logger.Log(devEUI, $"syncing FCntDown for multigateway", LogLevel.Information);
 
@@ -48,7 +48,7 @@ namespace LoRaWan.NetworkServer
             return 0;
         }
 
-        public override async Task<DeduplicationResult> CheckDuplicateMsgAsync(string devEUI, int fcntUp, string gatewayId, int fcntDown)
+        public override async Task<DeduplicationResult> CheckDuplicateMsgAsync(string devEUI, uint fcntUp, string gatewayId, uint fcntDown)
         {
             Logger.Log(devEUI, $"check for duplicate message", LogLevel.Debug);
